requirements.txt
amqp==2.5.2
asgiref==3.2.7
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.0
certifi==2019.11.28
# cffi==1.14.0
cffi==1.14.5
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
# cryptography==2.8
cryptography==3.4.6
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-rest-swagger==2.2.0
djangorestframework==3.11.2
# djangorestframework<4.0
# djangorestframework-simplejwt==4.4.0
djangorestframework-simplejwt==4.6.0
Flask==1.1.2
gunicorn==20.0.4
h11==0.9.0
# httptools==0.1.1
idna==2.9
importlib-metadata==1.5.2
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
# Pillow==7.0.0
Pillow==8.1.2
# psycopg2-binary==2.8.4
psycopg2-binary==2.8.6
pycparser==2.20
# PyJWT==1.7.1
PyJWT==2.0.1
python3-openid==3.1.0
# pytz==2019.3
pandas>=1.1.5
pytz==2021.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
# social-auth-app-django==3.1.0
social-auth-app-django==4.0.0
# social-auth-core==3.3.2
social-auth-core==4.1.0
soupsieve==2.0
sqlparse==0.3.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.11.3
# uvloop==0.14.0
vine==1.3.0
websockets==8.1
zipp==3.1.0
argon2-cffi==20.1.0
django-analytical==2.6.0
django-crispy-forms==1.9.1
# numpy==1.19.0
numpy==1.20.1
flask==1.1.2
drf-yasg==1.17.1
coverage==5.2.1
model-bakery==1.1.1
django-nose==1.4.7
nose==1.3.7
django-webtest==1.9.7
mock==4.0.2
django-fullurl==1.1
markdown==3.3.2
rest-social-auth==6.0.0
# Temporary packages in Django admin area 'django-admin-rangefilter'
django-admin-rangefilter>=0.7.0
scikit-learn==0.23.2
flaml==0.5.12
django-filter==2.4.0
